# go_backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smart Waste Management API
  description: |
    RESTful API for an AI & AR-Driven Smart Waste Management Platform.
    
    ## Features
    - **IoT Integration**: Real-time data from ESP32-based smart bins via MQTT
    - **User Management**: Profile and reward point tracking
    - **Driver Management**: Route optimization and task verification
    - **Dashboard**: Company management, pricing rules, and analytics
    - **Valuation Engine**: AI-based waste pricing
    
    ## Authentication
    Authentication endpoints are placeholders. Implement JWT or OAuth2 for production.
  version: 1.0.0
  contact:
    name: Smart Waste Team
    email: support@smartwaste.io
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    post:
      tags:
        - Users
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: Email already exists
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
    put:
      tags:
        - Users
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
    delete:
      tags:
        - Users
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
  /users/{id}/rewards:
    get:
      tags:
        - Users
      summary: Get user reward points
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reward points
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  reward_points:
                    type: integer
    post:
      tags:
        - Users
      summary: Add reward points
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRewardPointsRequest'
      responses:
        '200':
          description: Points added

  # Drivers
...
components:
  schemas:
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
        phone:
          type: string
        address:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: string
        address:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        phone:
          type: string
        address:
          type: string
        reward_points:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        meta:
          $ref: '#/components/schemas/Pagination'
    AddRewardPointsRequest:
      type: object
      required:
        - points
        - reason
      properties:
        points:
          type: integer
          minimum: 1
        reason:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer